Add testing for UnixStyleUsageFormatter
